+++ CHANGELOG.md
@@ -2,15 +2,22 @@
 
 ## Unreleased
 
+  - reconfigured priorities when low food alert is active
+  - injured prisoners will increase doctoring priority
+  - roaming animal alert will increase handling priority
+  - shooting frenzy increases desire to hunt
+  - deteriorating things in valid storage do not increase desire to haul
+  - colonists will now consider if they are the best at something and how their
+    skills rate relative to others
   - the Work Tab mod, if used, must be loaded after this one
   - colonists will no longer want to refuel items with autofuel disabled
-  - added a new mood buf for colonists who desire free will and have not had
+  - added a new mood buff for colonists who desire free will and have not had
     player forced jobs added for some time
   - reworked the precept impact level to more accurately show deviation from
     vanilla
   - factions without a free will precept will be given the "don't care" free
     will precept at the start of the game
-  - colonists are less likely to do unnecesary work when others are
+  - colonists are less likely to do unnecessary work when others are
     downed/injured
   - colonists who are downed, but in a bed, do not affect calculations of downed
     colonists